// ts/src/generated/accounts/marketStateHeader.ts
export type MarketStateHeader = {
  discriminator: ReadonlyUint8Array;
  marketId: Address;
  baseMint: Address;
  quoteMint: Address;
  baseVault: Address;
  quoteVault: Address;
  admin: Address;
  baseAtomsPerBaseLot: bigint;
  quoteAtomsPerQuoteLot: bigint;
  tickSizeInQuoteAtomsPerBaseUnit: bigint;
  rawBaseUnitsPerBaseUnit: bigint;
  uncollectedFeesQuoteLots: bigint;
  collectedFeesQuoteLots: bigint;
  makerFeePpm: number;
  takerFeePpm: number;
  baseDecimals: number;
  quoteDecimals: number;
  status: number;
  mode: number;
  marketBump: number;
  syncFeeMultiplier: number;
  minAsyncDelaySlots: number;
  maxAsyncDelaySlots: number;
  reserved: number;
};